requirements.txt
certifi==2020.12.5
chardet==4.0.0
DateTime==4.3
idna==2.10
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
urllib3==1.26.4
zope.interface==5.2.0
